+++ src/libraries/Common/src/Interop/OSX/Interop.Libraries.cs
@@ -9,6 +9,7 @@ internal static partial class Interop
         internal const string CoreFoundationLibrary = "/System/Library/Frameworks/CoreFoundation.framework/CoreFoundation";
         internal const string CoreServicesLibrary = "/System/Library/Frameworks/CoreServices.framework/CoreServices";
         internal const string CFNetworkLibrary = "/System/Library/Frameworks/CFNetwork.framework/CFNetwork";
+        internal const string libobjc = "/usr/lib/libobjc.dylib";
         internal const string libproc = "libproc";
         internal const string LibSystemCommonCrypto = "/usr/lib/system/libcommonCrypto";
         internal const string LibSystemKernel = "/usr/lib/system/libsystem_kernel";
@@ -16,6 +17,5 @@ internal static partial class Interop
         internal const string SystemConfigurationLibrary = "/System/Library/Frameworks/SystemConfiguration.framework/SystemConfiguration";
         internal const string AppleCryptoNative = "System.Security.Cryptography.Native.Apple";
         internal const string MsQuic = "msquic";
-
     }
 }
